+++ tests/unit_sync/conftest.py
@@ -10289,108 +10289,109 @@ def get_txevent_result():
         "jsonrpc": "2.0",
         "method": "sui_subscribeTransaction",
         "params": {
-            "subscription": 6037208326139487,
+            "subscription": 3209114665361176,
             "result": {
                 "certificate": {
-                    "transactionDigest": "DUwEegGA7kP3aLTtoauKJ414HJK6EeJiQGTvnCih6vYs",
+                    "transactionDigest": "9GsVP2JFc9FKKpMPqTY681CRnM1kqDcVwao33N9RJEdV",
                     "data": {
                         "transactions": [
                             {
                                 "Call": {
                                     "package": {
-                                        "objectId": "0x1c151df881200ec0d2381c2a6686983f69830991",
+                                        "objectId": "0x393b1b81a75697df6f21ceca589645e5d43d95bc",
                                         "version": 1,
-                                        "digest": "+SYUdKUDLaqeD1z5DiVeksSdJiHcN7nkcQ6h6T63xyc=",
+                                        "digest": "rmXi0Hqmu9nyCluQCXWiVuVWx2Imeh1mydYbJ0cG5nQ=",
                                     },
                                     "module": "marketplace_nofee",
                                     "function": "list",
                                     "typeArguments": [
-                                        "0x1c151df881200ec0d2381c2a6686983f69830991::meta_nft::MetaNFT",
+                                        "0x393b1b81a75697df6f21ceca589645e5d43d95bc::meta_nft::MetaNFT",
                                         "0x2::sui::SUI",
                                     ],
                                     "arguments": [
-                                        "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905",
-                                        "0x6c5428245ae9dda32ccf96d00de1fb228b09016",
-                                        10000000,
+                                        "0xa61e4919e1fabfa57cfee1433ccce6707eda956",
+                                        "0x5e435c99a0de0d865602415d47f214295fdfbd5b",
+                                        [128, 150, 152, 0, 0, 0, 0, 0],
                                     ],
                                 }
                             }
                         ],
                         "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
                         "gasPayment": {
-                            "objectId": "0x026fa7a636774249a723accde1b14b2ef48e398d",
-                            "version": 891211,
-                            "digest": "7Ur0lpcsX8m0Lfy0+FqGvU43bubAAO55XRGcQUcheRI=",
+                            "objectId": "0x1bc640fe42c46b94284969f77d51bcf1c88be168",
+                            "version": 127923,
+                            "digest": "9lVAV2AoGtjRHGOoucoeR005ao7Ddl2ZU5JpwqiE0qs=",
                         },
-                        "gasBudget": 2000,
+                        "gasPrice": 1,
+                        "gasBudget": 10000,
                     },
-                    "txSignature": "AMngKTI0zlxV2w4d3ndy50td49BU9IJkO7ZKpIG+8JJqjCld/LuGFI6R592/1nsH0uo7X4mm2epl23WciAVeDwKP2Q9j7jGq+XD9JhgVVHJZu6xHOLVLDXqSmiyXz2JAWw==",
+                    "txSignature": "AJqPpR06VO9xNtSNW4CcSU+ClMgHdjMZK7Q1sdToXhCTzehLqSk6bqhoUkNQbnQPcXXgjKk4eIdl1TTFe55U4QiP2Q9j7jGq+XD9JhgVVHJZu6xHOLVLDXqSmiyXz2JAWw==",
                     "authSignInfo": {
-                        "epoch": 0,
-                        "signature": "q/zstQjxM89sN7MGqPzzwUt0KV8UJoLJlL0zQIj3ztplFRRKZqFol0NBUI1+784G",
-                        "signers_map": [58, 48, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 16, 0, 0, 0, 0, 0, 1, 0, 3, 0],
+                        "epoch": 5,
+                        "signature": "AaHaRRgm8pGyqgxmCIEUfe4jUvvpG32a0Php4sMhsLPdvP44L85NkIc80xCJGo69sw==",
+                        "signers_map": [58, 48, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 16, 0, 0, 0, 1, 0, 2, 0, 3, 0],
                     },
                 },
                 "effects": {
                     "status": {"status": "success"},
-                    "gasUsed": {"computationCost": 377, "storageCost": 99, "storageRebate": 59},
+                    "gasUsed": {"computationCost": 380, "storageCost": 97, "storageRebate": 56},
                     "sharedObjects": [
                         {
-                            "objectId": "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905",
-                            "version": 892629,
-                            "digest": "/9Dbd2WKGR0Qd/eQ4Vjrph/f2FZ341/i8UHq7/cd2pw=",
+                            "objectId": "0x0a61e4919e1fabfa57cfee1433ccce6707eda956",
+                            "version": 127922,
+                            "digest": "WGrYA/okuXtaTQJA5QxaT8WcQraSN0wvgd8FM7wm6JY=",
                         }
                     ],
-                    "transactionDigest": "DUwEegGA7kP3aLTtoauKJ414HJK6EeJiQGTvnCih6vYs",
+                    "transactionDigest": "9GsVP2JFc9FKKpMPqTY681CRnM1kqDcVwao33N9RJEdV",
                     "created": [
                         {
-                            "owner": {"ObjectOwner": "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905"},
+                            "owner": {"ObjectOwner": "0x94cca84604ab8892c529f0163ce233de275c2efd"},
                             "reference": {
-                                "objectId": "0x64e861f251e710585b10d7ae1cd3f0b709bc3d88",
-                                "version": 892630,
-                                "digest": "s2JOCsTxLe3Oxm7jcIsfx/byEy2CKtY6zYOZIg4LhDc=",
+                                "objectId": "0x24c339ca58e2469b2116e2819a25bf69e5e4724e",
+                                "version": 127924,
+                                "digest": "EW8WCaC026uERnxaP+sMWwAqmKW6xcvbC1600s+RmMs=",
                             },
                         },
                         {
-                            "owner": {"ObjectOwner": "0x64e861f251e710585b10d7ae1cd3f0b709bc3d88"},
+                            "owner": {"ObjectOwner": "0x0a61e4919e1fabfa57cfee1433ccce6707eda956"},
                             "reference": {
-                                "objectId": "0x9b2b86ebf7dbb3571dcf526d593ebba0caab4c9a",
-                                "version": 892630,
-                                "digest": "YRKCuivYTuIjoNP1UAuT/q9xOM0zMhCqOPQJ8VtLnm8=",
+                                "objectId": "0x94cca84604ab8892c529f0163ce233de275c2efd",
+                                "version": 127924,
+                                "digest": "e6e0zuqNYUwl6xee9nwNnLfYqExTeZocqN5yC3OgWPk=",
                             },
                         },
                     ],
                     "mutated": [
                         {
-                            "owner": {"AddressOwner": "0x15b49685f5360c83e930410bb266896fcb676b77"},
+                            "owner": {"Shared": {"initial_shared_version": 44}},
                             "reference": {
-                                "objectId": "0x026fa7a636774249a723accde1b14b2ef48e398d",
-                                "version": 892630,
-                                "digest": "nheEtAXfWJZEntpaBR/arxCWSJWU14PaeajDW251D6o=",
+                                "objectId": "0x0a61e4919e1fabfa57cfee1433ccce6707eda956",
+                                "version": 127924,
+                                "digest": "odRipsnD5bwu1PBeoq69jJUnhE+JccJTB5uGZURXqos=",
                             },
                         },
                         {
-                            "owner": {"Shared": {"initial_shared_version": 464}},
+                            "owner": {"AddressOwner": "0x15b49685f5360c83e930410bb266896fcb676b77"},
                             "reference": {
-                                "objectId": "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905",
-                                "version": 892630,
-                                "digest": "Fm6u8U6ynSqH3Zh63joIlSJhz+hhGid8an6/Monq33k=",
+                                "objectId": "0x1bc640fe42c46b94284969f77d51bcf1c88be168",
+                                "version": 127924,
+                                "digest": "K5ppNIXvOwFT4f0Ajjxrudxfj0ZNZ6fLssVi5O4Rzco=",
                             },
                         },
                     ],
                     "wrapped": [
                         {
-                            "objectId": "0x06c5428245ae9dda32ccf96d00de1fb228b09016",
-                            "version": 892630,
+                            "objectId": "0x5e435c99a0de0d865602415d47f214295fdfbd5b",
+                            "version": 127924,
                             "digest": "WFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFg=",
                         }
                     ],
                     "gasObject": {
                         "owner": {"AddressOwner": "0x15b49685f5360c83e930410bb266896fcb676b77"},
                         "reference": {
-                            "objectId": "0x026fa7a636774249a723accde1b14b2ef48e398d",
-                            "version": 892630,
-                            "digest": "nheEtAXfWJZEntpaBR/arxCWSJWU14PaeajDW251D6o=",
+                            "objectId": "0x1bc640fe42c46b94284969f77d51bcf1c88be168",
+                            "version": 127924,
+                            "digest": "K5ppNIXvOwFT4f0Ajjxrudxfj0ZNZ6fLssVi5O4Rzco=",
                         },
                     },
                     "events": [
@@ -10402,60 +10403,60 @@ def get_txevent_result():
                                 "changeType": "Gas",
                                 "owner": {"AddressOwner": "0x15b49685f5360c83e930410bb266896fcb676b77"},
                                 "coinType": "0x2::sui::SUI",
-                                "coinObjectId": "0x026fa7a636774249a723accde1b14b2ef48e398d",
-                                "version": 891211,
-                                "amount": -417,
-                            }
-                        },
-                        {
-                            "newObject": {
-                                "packageId": "0x1c151df881200ec0d2381c2a6686983f69830991",
-                                "transactionModule": "marketplace_nofee",
-                                "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
-                                "recipient": {"ObjectOwner": "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905"},
-                                "objectType": "0x2::dynamic_field::Field<0x2::dynamic_object_field::Wrapper<0x2::object::ID>, 0x2::object::ID>",
-                                "objectId": "0x64e861f251e710585b10d7ae1cd3f0b709bc3d88",
-                                "version": 892630,
-                            }
-                        },
-                        {
-                            "newObject": {
-                                "packageId": "0x1c151df881200ec0d2381c2a6686983f69830991",
-                                "transactionModule": "marketplace_nofee",
-                                "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
-                                "recipient": {"ObjectOwner": "0x64e861f251e710585b10d7ae1cd3f0b709bc3d88"},
-                                "objectType": "0x1c151df881200ec0d2381c2a6686983f69830991::marketplace_nofee::Listing<0x1c151df881200ec0d2381c2a6686983f69830991::meta_nft::MetaNFT, 0x2::sui::SUI>",
-                                "objectId": "0x9b2b86ebf7dbb3571dcf526d593ebba0caab4c9a",
-                                "version": 892630,
+                                "coinObjectId": "0x1bc640fe42c46b94284969f77d51bcf1c88be168",
+                                "version": 127923,
+                                "amount": -421,
                             }
                         },
                         {
                             "mutateObject": {
-                                "packageId": "0x1c151df881200ec0d2381c2a6686983f69830991",
+                                "packageId": "0x393b1b81a75697df6f21ceca589645e5d43d95bc",
                                 "transactionModule": "marketplace_nofee",
                                 "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
-                                "objectType": "0x1c151df881200ec0d2381c2a6686983f69830991::marketplace_nofee::Marketplace",
-                                "objectId": "0xdcc4acdcbe40459d37e0ff1d5518cd38705ea905",
-                                "version": 892630,
+                                "objectType": "0x393b1b81a75697df6f21ceca589645e5d43d95bc::marketplace_nofee::Marketplace",
+                                "objectId": "0x0a61e4919e1fabfa57cfee1433ccce6707eda956",
+                                "version": 127924,
+                            }
+                        },
+                        {
+                            "newObject": {
+                                "packageId": "0x393b1b81a75697df6f21ceca589645e5d43d95bc",
+                                "transactionModule": "marketplace_nofee",
+                                "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
+                                "recipient": {"ObjectOwner": "0x94cca84604ab8892c529f0163ce233de275c2efd"},
+                                "objectType": "0x393b1b81a75697df6f21ceca589645e5d43d95bc::marketplace_nofee::Listing<0x393b1b81a75697df6f21ceca589645e5d43d95bc::meta_nft::MetaNFT, 0x2::sui::SUI>",
+                                "objectId": "0x24c339ca58e2469b2116e2819a25bf69e5e4724e",
+                                "version": 127924,
+                            }
+                        },
+                        {
+                            "newObject": {
+                                "packageId": "0x393b1b81a75697df6f21ceca589645e5d43d95bc",
+                                "transactionModule": "marketplace_nofee",
+                                "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
+                                "recipient": {"ObjectOwner": "0x0a61e4919e1fabfa57cfee1433ccce6707eda956"},
+                                "objectType": "0x2::dynamic_field::Field<0x2::dynamic_object_field::Wrapper<0x2::object::ID>, 0x2::object::ID>",
+                                "objectId": "0x94cca84604ab8892c529f0163ce233de275c2efd",
+                                "version": 127924,
                             }
                         },
                         {
                             "deleteObject": {
-                                "packageId": "0x1c151df881200ec0d2381c2a6686983f69830991",
+                                "packageId": "0x393b1b81a75697df6f21ceca589645e5d43d95bc",
                                 "transactionModule": "marketplace_nofee",
                                 "sender": "0x15b49685f5360c83e930410bb266896fcb676b77",
-                                "objectId": "0x06c5428245ae9dda32ccf96d00de1fb228b09016",
-                                "version": 892630,
+                                "objectId": "0x5e435c99a0de0d865602415d47f214295fdfbd5b",
+                                "version": 127924,
                             }
                         },
                     ],
                     "dependencies": [
-                        "5GCJXSMJVzNXJ8mwx2UhUpcGb4NEVGPqu9436srUEbTx",
-                        "BTkzNcW1uzUtjE4jYgMTco1mstLxAMa24M9syQEoxkmb",
-                        "CgoxqazRuZpxnU5WoF5UxE9TDHE7Yv2aDEwykEMBmtMP",
+                        "8VRBCQuDzKjsSiDGJdRWQxACnnN6T9mD5Pb1SvmdufPN",
+                        "AcFwWyMEJzxuG3pNcXSdFAktq2SkWUhpPCE58NctyQmM",
+                        "CuzekBAGTEHoBT3d6x1MDG8wJafLKqtxsGU8D4k7cBxo",
                     ],
                 },
-                "timestamp_ms": 1672309762188,
+                "timestamp_ms": 1674216639927,
                 "parsed_data": None,
             },
         },